+++ trax/Cell.h
@@ -9,10 +9,10 @@ struct Cell
   Cell() = delete;
 
   // clang-format off
-  Cell(double X1, double Y1, double Z1,
-       double X2, double Y2, double Z2,
-       double X3, double Y3, double Z3,
-       double X4, double Y4, double Z4,
+  Cell(float X1, float Y1, float Z1,
+       float X2, float Y2, float Z2,
+       float X3, float Y3, float Z3,
+       float X4, float Y4, float Z4,
        int I, int J)
       : x1(X1), y1(Y1), z1(Z1),
         x2(X2), y2(Y2), z2(Z2),
@@ -23,11 +23,11 @@ struct Cell
   }
   // clang-format on
 
-  double x1, y1, z1;  // below left
-  double x2, y2, z2;  // above left
-  double x3, y3, z3;  // above right
-  double x4, y4, z4;  // below right
-  int i;              // below left corner indices
+  float x1, y1, z1;  // below left
+  float x2, y2, z2;  // above left
+  float x3, y3, z3;  // above right
+  float x4, y4, z4;  // below right
+  int i;             // below left corner indices
   int j;
 };
 
@@ -38,7 +38,7 @@ bool is_saddle(const Cell& cell);
 bool first_diagonal_larger(const Cell& cell);
 
 // Return minmax values of a grid cell
-using MinMax = std::pair<double, double>;
+using MinMax = std::pair<float, float>;
 MinMax minmax(const Cell& cell);
 
 }  // namespace Trax
